# daveslist-rnqjne/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/sell_car/car_list": {
            "get": {
                "tags": [
                    "car_list"
                ],
                "operationId": "get_sell_car",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/responses.CarCategoriesResp"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.UnauthorizedErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new car listing with associated files",
                "tags": [
                    "car_list"
                ],
                "summary": "Create a new car listing",
                "operationId": "create_car_list",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Title",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Content",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "file"
                        },
                        "collectionFormat": "csv",
                        "description": "Uploaded files",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Is the listing public?",
                        "name": "is_public",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.NoDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.UnauthorizedErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/sell_car/car_list/{id}": {
            "put": {
                "tags": [
                    "car_list"
                ],
                "operationId": "edit_car_list",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "selling car data",
                        "name": "selling",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RequestSellingCar"
                        }
                    },
                    {
                        "type": "string",
                        "description": "selling car id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.NoDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.UnauthorizedErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "reply"
                ],
                "operationId": "reply_sell_car",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "reply data",
                        "name": "reply",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.Reply"
                        }
                    },
                    {
                        "type": "string",
                        "description": "selling car id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.NoDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.UnauthorizedErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "car_list"
                ],
                "operationId": "delete_car_list",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "selling car id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.NoDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.UnauthorizedErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/sell_car/car_list/{id}/hide": {
            "put": {
                "tags": [
                    "car_list"
                ],
                "operationId": "hide_car_list",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "want to hide or not",
                        "name": "hide",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.RequestHideList"
                        }
                    },
                    {
                        "type": "string",
                        "description": "car_list id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.NoDataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.BadRequestErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.UnauthorizedErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.Reply": {
            "type": "object",
            "required": [
                "text"
            ],
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
...
        "requests.RequestHideList": {
            "type": "object",
            "properties": {
                "is_show": {
                    "type": "boolean"
                }
            }
        },
        "requests.RequestSellingCar": {
            "type": "object"
        },
        "responses.BadRequestErrorResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean",
                    "x-order": "0",
                    "example": false
                },
                "code": {
                    "type": "integer",
                    "x-order": "1",
                    "example": 400
                },
                "error": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "x-order": "0",
                            "example": "bad request error message"
                        },
                        "field": {
                            "type": "object",
                            "x-order": "1"
                        }
                    },
                    "x-order": "2"
                }
            }
        },
        "responses.CarCategoriesResp": {
            "type": "object",
            "properties": {
                "car_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.ListResp"
                    }
                },
                "category_name": {
                    "type": "string"
                }
            }
        },
        "responses.FileResp": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "responses.InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean",
                    "x-order": "0",
                    "example": false
                },
                "code": {
                    "type": "integer",
                    "x-order": "1",
                    "example": 500
                },
                "error": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "x-order": "0",
                            "example": "internal server error"
                        },
                        "field": {
                            "type": "object",
                            "x-order": "1"
                        }
                    },
                    "x-order": "2"
                }
            }
        },
        "responses.ListResp": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "file_img": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.FileResp"
                    }
                },
                "reply": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.ReplyResp"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "responses.NoData": {
            "type": "object"
        },
        "responses.NoDataResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean",
                    "x-order": "0",
                    "example": true
                },
                "code": {
                    "type": "integer",
                    "x-order": "1"
                },
                "data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/responses.NoData"
                        }
                    ],
                    "x-order": "2"
                }
            }
        },
        "responses.ReplyResp": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        },
        "responses.Success": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "status": {
                    "type": "boolean"
                }
            }
        },
...
        "responses.UnauthorizedErrorResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean",
                    "x-order": "0",
                    "example": false
                },
                "code": {
                    "type": "integer",
                    "x-order": "1",
                    "example": 401
                },
                "error": {
                    "type": "object",
                    "properties": {
                        "message": {
                            "type": "string",
                            "x-order": "0",
                            "example": "unauthorized to access the resource"
                        },
                        "field": {
                            "type": "object",
                            "x-order": "1"
                        }
                    },
                    "x-order": "2"
                }
            }
        },
...
    }
}